Split endpoints to two folders and return mock data
